+++ script/lib/Deployer.sol
@@ -12,6 +12,7 @@ import {LocalFork} from "@mgv/test/lib/forks/Local.sol";
 import {TestnetZkevmFork} from "@mgv/test/lib/forks/TestnetZkevm.sol";
 import {GoerliFork} from "@mgv/test/lib/forks/Goerli.sol";
 import {SepoliaFork} from "@mgv/test/lib/forks/Sepolia.sol";
+import {BlastSepoliaFork} from "@mgv/test/lib/forks/BlastSepolia.sol";
 import {ZkevmFork} from "@mgv/test/lib/forks/Zkevm.sol";
 import {console2 as console} from "@mgv/forge-std/console2.sol";
 
@@ -71,6 +72,8 @@ abstract contract Deployer is Script2 {
         fork = new MumbaiFork();
       } else if (block.chainid == 11155111) {
         fork = new SepoliaFork();
+      } else if (block.chainid == 168587773) {
+        fork = new BlastSepoliaFork();
       } else {
         revert(string.concat("Unknown chain id ", vm.toString(block.chainid), ", cannot deploy."));
       }