+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Upgrade to @mangrovedao/mangrove-deployments v2.0.1
 - Upgrade to @mangrovedao/context-addresses v1.1.0
+- Simplify copying of context addresses
 
 # 2.0.3