+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Next version
 
 - Upgrade to @mangrovedao/mangrove-deployments v2.0.1
-- Upgrade to @mangrovedao/context-addresses v1.1.0
+- Upgrade to @mangrovedao/context-addresses v1.1.2
 - Simplify copying of context addresses
+- Add support for Blast Sepolia
 
 # 2.0.3