+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+# 2.1.0-0
+
 - Upgrade to @mangrovedao/mangrove-deployments v2.0.1
 - Upgrade to @mangrovedao/context-addresses v1.1.2
 - Simplify copying of context addresses